Add tests for bug 1041 SON (option scale)
I previously added 2 tests for this issue, for the bsim3 and bsim4.

This commit adds tests for the MOS1, MOS2, MOS3, MOS6, B3SOI and VDMOS.
It also renames the original bsim3 and bsim4 tests to indicate what
devices they are for, and that they are "scale" tests.

All of these tests follow the same pattern.  Two versions of a circuit
are run, that are identical other than the use of

.options parser scale=1.0e-6

Then the results of these two runs are compared.