Tiny update, changing case on one gold std msg
This is related to Issue #101.

For various reasons, I changed how case is handled in the new expression
library to support reading of files.  I can't up-case the entire
expression string anymore, as filenames need to preserve case on
linux file systems.

One result is that error messages sometimes have a different case now.
That appears to have broken a single error test.  This commit fixes
the issue.