Test Rise/Fall/Cross Vals < -1 for Extrema Measures
This commit tests that RISE, FALL and CROSS values less than -1
default to -1, with the appropriate warning message, for the MAX,
MIN and PP measures.  This commit also updated some of the MIN
measure tests to use V(2) so that there is more than one fall in
the measured waveform.

SON Bug 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313