Update two transmission line gold standards
These two gold standards were generated long ago and we never noticed
that they contained some strange interpolation artifacts close to
discontinuities.

When interpolation in the transmission line was modified to get rid of
other artifacts in circuits of customer interest, these two tests
began to fail.  In both cases it was because Xyce is no longer
producing the artifacts that are present in the gold standard, and the
test is complaining.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/103