Updates scripts to run properly when used with MPI-prefixed binary
If the Xyce binary being passed in is prefixed by "mpirun -np <#>",
then the shell script does not find the correct $XYCEROOT.  The logic
was improved to find the correct $XYCEROOT in the shell script of
Certification_Tests/XyceAsLibrary/BUG_1685/c7i.cir.sh when run with
MPI.  This was not migrated to the other scripts because these tests
are not run in parallel and the parallel build - serial execution
testing does not prefix the Xyce binary with "mpirun".  However,
in the event that this policy changes, or someone accidentally
(that would be me) runs the parallel build - serial execution tests
with "mpirun", then it doesn't fail.