Adds corrected block Jacobi test for invert_hb
The HB block jacobi preconditioner has a one-step correction option
that has been implemented for some time.  However, this correction has
not been tested nightly, which is not helpful when updating linear
algebra infrastructure and looking at code coverage.  This test is
being added to the HB inverter test because it is one of the circuits
where this preconditioner does reduce the number of linear solver
iterations.