Allow Regression Tests to Use Files in Subdirectories
This commit allows Xyce regresson tests to use files that are located
in subdirectories of the directory that contains the "top-level"
netlist.  It is in proximate support of SON Bug 1325.  The constraints
on this new capability are:

  1) All of the files needed for the test Test_Dir/top should be in
the /path/to/Xyce_Regression/Netlists/Test_Dir/Manifest.txt file.

  2) If the file /path/to/Xyce_Regression/Netlists/Test_Dir/sub1/inc.cir
used for the test Test_dir/top ends in .cir, then it needs to be in
the /path/to/Xyce_Regression/Netlists/Test_Dir/sub1/exclude file.

  3) If no other tests are run from the sub1 subdirectory then the
only tags file is /path/to/Xyce_Regression/Netlists/Test_Dir/tags.

  4) The use of . and .. characters is not supported in the relative
paths for the Test_Dir/top files in subdirectories.  In addition,
since the entire Xyce_Regression test suite is very "Linux-centric",
only the '/' character is supported as a file separator in the
relative paths.

A concrete example will be given by the files for testing SON Bug 1325.

SON Bugs 1325 and 1326
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1325
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1326