Initial Tests for SON Bug 1325 (INC File Paths)
This is the initial set of tests for SON Bug 1325.  It tests the
"baseline" cases that work in Xyce 7.2, and earlier, for .INC files.
It covers the two cases of:

  1) an absolute path on the .INC line.

  2) a path on a .INC line in a file (e.g., sub1/include1) that is
relative to the directory of the top-level netlist rather than to
the directory (sub1) of the containing file.

SON Bug 1325
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1325