Disable "rxn" tests if binary doesn't have the parser
It is possible to disable compilation of the reaction parser in
Xyce (but it is rarely done).

Until the incorporation of the new expression package, the main reason
for doing this was so that one did not need flex and bison.

We now need flex and bison anyway.  It is still possible to turn off
the reaction parser, but one must then not run any of the reaction
parser's tests.

Most of those tests are in the OUO repo and are only enabled when the
parser is available.  But there are some in the open repo that don't
have "required" tags.  This commit adds "-rxn" to a tags list if the
binary given to suggestXyceTagList.sh doesn't have the reaction parser
built in.

This is inspired by manually testing issue #113, but is not required
for that fix.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/113