Add Sub-Tests of Case Sensitivity for S, Y and Z Ops
This commit changes various usages of S(), Y() and Z() operators,
in existing regression tests, to use the lower-case versions.  This
was motivated by Eric's recent changes to the unit tests for the
expression library.  (See commit 22b7add1 from 12/2/2020.)  All of
these tests pass, with the lower-case versions, with the Xyce 7.2
release.  So, there was no issue in Xyce 7.2.