Add more AC Sens tests, especially for nonlinear params
When I implemented AC sensitivities last year I didn't create very many
tests.  The tests only test instance parameters in linear devices such as
the capacitor, where it was easy to create analytic matrix derivatives.

For context; the AC sensitivities calculation requires a matrix
derivative on the RHS of the sensitivity linear system, that I call J'.
This is the derivative of the AC Jacobian w.r.t. the sensitivity
parameter.

Many devices (anything nonlinear) need to use numerical matrix
derivatives.  For that I created functions to compute them,
this use case wasn't covered in regression testing at all.

This commit corrects this mistake, by testing several aspects of
numerical matrix derivatives.

This work was prompted by a recent user report, that he was getting seg
faults when attempting to compute AC sensitivities for MOSFET L, W parameters.
The seg faults were happening because I had an indexing mistake in the
aforementioned numerical matrix derivatives function.   There was also a
math mistake, which I have also corrected.  This is documented in gitlab
Issue #115.

Here are some details about some of the new tests:

gain-stage.cir - this is a bsim3 test, and the .sens param is the L,W of
the MOSFET device.  It is compared to a finite difference.
The freqency range on the gain stage test circuit is pretty small,
only from 1e7 to 1e9.  This is b/c these sensitivities (of voltage node
2 w.r.t. m1:w and m1:l) are both really tiny until the frequency is
high.

op-amp.cir tests - this is also a bsim3 test, compared to a finite difference.
There are actually 2 tests of the op-amp in this commit.  The first (original)
one is op-amp.cir, which is compared to op-amp_fd.cir.  The second one is
op-amp_unique.cir, which is compared to op-amp_fd_unique.cir.
They are basically the same test, but I had a theory that numerical AC
sensitivities was broken when the same node was attached to more than one
terminal of a device.    So, this circuit tests that theory.

lowpass_forceFD*.cir - Several other tests are added to ensure that the the
numerical calculation of J' (derivative of the AC Jacobian w.r.t.  a
parameter) matches the analtical answer.  This can only be tested for
linear devices.  For these 3 tests, I am running the same circuit twice,
once with analytical derivatives and once with numerical derivatives, and
then I compare the results.