Excluding the mutind1 test in this directory
This is related to bug 1190 on the SON bugzilla.

This test relies on file_compare, which means that it fails if the
mutind1.cir.prn and mutind1_baseline.cir.prn files have even slightly
different numbers of time steps.

I wrote it that way (instead of  using the more usual xyce_verify.pl)
because this is supposed to be a test that runs the same problem 2x,
with the only difference being that the netlists have slightly different
methods for setting up the problem.    Given that, using "diff" arguably
could have worked, and using file_compare was an attempt to make this a
slightly safer test.

This test has been passing everywhere until recently, when it started
failing in with the new trilinos-dev builds.  It isn't at all clear to
me why swapping the trilinos version would cause this to fail, but we
are where we are.

Bug 1190 is still in the "assigned" state so, it isn't a fixed bug
anyway.  I plan to make a note of this issue in the comments for that
bug.  In the meantime I am exlcuding this test, until I can fix it
properly.