Exclude BUG_1190_SON/mutInd2
We started noticing random failures of BUG_1190_SON/mutInd1 the past
two weeks, and initially attributed it to the one special build in
which both failures happened.

Turns out that it has nothing to do with that build, and is just a
rare, random failure that fails in parallel about 1% of the time.  It
can be made to fail much more often by running on more than 2 processors.

It also turns out that mutInd2 is *also* subject to the same random
failures, but only about 0.2% of the time (on 2 processors, much more
frequently on more than 2).

Since the issue that is meant to be tested here is not in fact
completely fixed anyway *and* the test will randomly fail for reasons
completely unrelated to the issue it is meant to test, I'm disabling
mutInd2 just the way that mutInd1 was excluded yesterday.

http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/116