Removing ACSENS/bsim3/gain-stage from testing
This test is only failing on Windows in regression testing, but as I
won't have time to diagnose and fix the issue today, I am removing it
from testing.

As with the ACSENS/bsim3/op-amp tests, I am guessing this is failing due
to limitations of the AC comparator, combined with the fact that these
tests all have zero crossings and really large magnitudes.   Finally,
performing finite differences on AC calculations often seems to produce
noisy data, which is another difficulty with this sort of test.