Fix ylin-2port-yparam-matrixFormat for Parallel Failures
This test was failing in parallel, on ascic180 on RHEL7, when run
with 3 or 4 processors.  It was not failing on any platform during
nightly testing.  This commit changes the comparator, for the
"non-base cases" of the Upper and Lower keywords on the [Matrix Format]
line in the Touchstone 2 input file, from diff to file_compare.pl.
While the output  for all three cases (including the Full base case)
should be identical in serial, they may differ slightly, in the
"numerically zero" values in their output, for parallel runs.  The
use of file_compare.pl with the specified tolerances should still
be sensitive enough to catch parsing errors during the YLIN device
read-in and setup.

SON Bug 1239
http://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1239