Modify C_T/BUG_966_SON Tests for Parallel Failures
This commit changes two tests in Certification_Tests/BUG_966_SON to
account for failures during 3- and 4-processor parallel testing.
The ordering of the output variables in the RAW file output differed
in some runs.  The .sh files for those two tests now use the same
approach as other tests in that subdirectory, such as tran-homotopy,
rather than using a simple diff of the filtered ("sans the date line")
RAW file output.

SON Bug 1239
http://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1239