Improve MIXED_SIGNAL/VPI Coverage of ADC/DAC Methods
This commit adds two new tests to MIXED_SIGNAL/VPI.  The purpose is
to cover more of the XyceCInterface functions related to the YADC and
YDAC devices.  This is a stepping stone to adding valgrind testing
of the Mixed Signal Interface functions, which will be the subject
of a TBD git issue.

SON Bugs 1067 and 1069
http://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1067
http://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1069