Add two new POLY tests for higher order
This commit adds the following two POLY tests:

(1) 1-input, 4th order
(2) 2-input, 3rd order

These tests also feature the use of POLY with .param-specified
coefficients, and also inside of a .func.  Prior to the recent work on
POLY, neither of these would have worked with the new expression
library.

Both of these tests contain (essentially) 2 circuits, one using POLY and the
other using an equivalent conventional polynomial expression.

For the first test (1-input, 4th order), on the .PRINT line, the output of
one "circuit" is subtracted from the other, and then that difference is
subtracted from 100.  So, every entry in the output file *should* equal 100.

For the second test, I don't explicitly compare the two outputs,
I just send both outputs to the *prn file.  That works fine as well
That works fine as well.