Initial Set of Tests for .FFT Lines
This commit adds the initial set of tests for .FFT lines.  There
are two error handling tests for invalid .FFT lines.

The test FFT/WindowTest does NOT produce the correct output yet.
Its main purpose is to enable valgrind testing for this new (and
not yet functional) .FFT capability.  It also tests parsing of the
NP and WINDOW qualifiers on .FFT lines, and ensures that multiple
FFT lines can be parsed.

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327