Removes normlvl from the list of nonlinear solver options in netlists
The "normlvl" option in the nonlinear solver package is effectively
ignored by NOX and not consistently used in Xyce's native Newton
solver.  It is the norm that is supposedly used by the nonlinear
solver, but NOX only recognizes normlvl=2.  To simplify the user
interface and code, this option will be removed in a cursory
cleanup of the nonlinear solver that will be committed soon.