Improve Tests for .FFT
This commit improves the tests for .FFT lines.  It now tests the
HAMM and HANN windows for FORMAT=NORM and FORMAT=UNORM, with the
more complete output produced by .OPTIONS FFT FFTOUT=1.  It also
tests that I(R1) works on a .FFT line even if I(R1) does not
appear on the .PRINT TRAN line.  It also tests expressions on the
.FFT line.  Case sensitivity for the FORMAT and WINDOW qualifiers
is also tested.

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327