Improve Tests of FFT Interface with Even Num. of Samples
This commit adds two subtests to FFT/WindowTest.  The purpose is to
improve testing of the FFT Interface with an even number of samples.
These subtests would give the incorrect answer for the magnitude
and phase of the 4th harmonic before commit a96114b to the Xyce
repo on 1/19/2021.

SON Bug 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327