Add Initial Set of Tests for .MEASURE FFT
This commit adds the initial set of tests for .MEASURE FFT.  It
covers the ENOB, SFDR, SNDR and THD measure types.  It covers the
use of voltage difference operators, expressions and lead currents
on .MEASURE FFT lines.  It also covers the error handling if the
.TRAN netlist does not have any .FFT lines, or lacks a .FFT line
with the same output variable as a given .MEASURE FFT line.

This commit also updates the FFT/WindowTest test to include a
voltage difference operator.  This was done because the MEASURE_FFT
tests assume that operator works correctly on .FFT lines.

SON Bugs 1280 and 1327
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1327