Test NBHARM Qualifier for THD for .MEASURE FFT
This commit adds subtests for the NBHARM qualifier for the THD
measure for .MEASURE FFT.  The results for the "special cases" match
HSPICE.  So, NBHARM=1 returns the value of the noiseFloor_
(which is -200db).  If NBHARM <= 0 or NBHARM > NP/2 (where NP is
the number of points in the FFT), the THD value using all of the
harmonics is returned.  Finally, the THD is always returned in dB,
for .MEASURE FFT.