Test .STEP with .FFT and .MEASURE FFT
This commit changes the test FFT/fft_accurateOptionZero so that
it should pass on all supported platforms, modulo some tweaking
of the comparator tolerances.  If that test works, then that
netlist is used to "bootstrap" testing of .STEP with .FFT and
then .STEP with .MEASURE FFT.  Those last two test sets assume
that the non-step cases are correct and just verify that the
.STEP tests get the same answers as the two separate non-step
netlists.  Both of the .OPTIONS FFT FFT_ACCURATE settings must
be tested for .STEP with .FFT, to ensure that breakpoint insertion
and interpolation for .FFT still work correctly with .STEP.  It
is then sufficient to just test FFT_ACCURATE=1 for .STEP with
.MEASURE_FFT.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327