Clean Up .sh Files for MEASURE_DC and MEASURE Tests
This commit cleans up various .sh files for tests in MEASURE_DC and
MEASURE.  The wrong measure file extension (mt0 instead of ms0) was
used in the .sh files for the MEASURE_DC tests that supported
valgrind testing.  This had no effect on the pass/fail correctness
of those tests, but it would produce less useful diagnostic output
if the -remeasure portion of the valgrind test failed.