Test -remeasure for .MEASURE FFT
This commit adds tests for -remeasure when the netlist has both
TRAN and FFT mode measures in it.  It covers both the step and
non-step cases.  Since the output from the .FFT lines is re-output
by -remeasure, the correctness of both the <netlistName>.mtx and
<netlistName>.fftx files is verified.

The window function for the first .FFT V(1) line was also changed
from RECT to BART.  The RECT window outputs nans (or infs) for
some of the metrics in the FFT output, which does not play well
with the file_compare.pl comparator.  I verified that the results
for the BART window are sufficiently different from the results
for the HANN window that the fact that the second .FFT V(1) is
used by the .MEASURE FFT lines is still verified.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327