Test Use of FFT-Based Measures in EQN/PARAM Measures
This commit adds subtests for the use of ENOB, FIND and THD measures,
which are "FFT-based" measures, in both TRAN and FFT mode EQN and
PARAM measures.  It tests that usage both within and outside of
an expression context.  Additional tests for the SFDR, SNR and
SNDR measure types are needed.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1330