Add clues to output produced by erroroption test
TIA/ERROROPTION/erroroption is a goofy little test that tries to check
that then-new "ERROROPTION" settings actually worked exactly as
intended.

It did so by relying on very verbose output from the time
integrator (by which I mean "all the stuff in TimeIntegrationPKG").
The test only works when Xyce is built with "--enable-verbose_time
--enable-verbose_nonlinear --enable-verbose_linear", so it is tagged
with "required:verbose".

This test started failing after commit 076d26678, which changed
breakpoint handling and time step selection logic.

It turns out that it's failing because it's not seeing time step
changes it's expecting.  That is almost certainly because the verbose
output is now out of sync with the changes in time step selection
logic, printing information before it's changed (or something).

It took me a long time to figure out that this was what was going on
precisely because there are several lines of code that set "failed=1"
in the perl script here but that emit messages from multiple test
stages to a file of verbose output without prefixing it with anything
indicating it's special, and without delimiting different test runs in
that file.

The standard output of the test itself is even less helpful, simply
saying stuff like "Failed 1" and "Failed 2".

This commit adds "FAILURE: " prefixes to every line of output that
happens whenever failed is set to 1.  That was present already on SOME
of them, but not the ones that were failing this time.

It also adds to the output the name of the netlist that is problematic
--- these are temporary netlists that have been produced by the
script and have somewhat randomized names, so it sorta helps to know
the actual file that had the problem, yeah?