Fix typo introduced in prior commit
In my prior commit, intended only to add "clues" to erroroption test
output, I fat-fingered and inserted a "p" into a variable name that
was important.  As a result, a failure condition that had previously
been detected was now being ignored and the test started passing
incorrectly.

I'm fixing the glitch, which means TIA/ERROROPTION/erroroption is
failing again, because it's a dumb, fragile test that probes internals
of the code via verbose output statements that can be out of sync with
the script's expectations.  This test and others similar to it that
depend on verbose output should be removed.