More Tests of FFT-Based Measures in EQN/PARAM Measures
This commit adds subtests for the use of SFDR, SNDR and SNR measures,
which are "FFT-based" measures, in FFT mode PARAM measures.  It only
tests that usage within an expression context, since the use of
measure names that start with 'S' do not work (yet), as the objective
variable, outside of an expression context in EQN/PARAM measures.
(This is because of issues with the parsing in extractMEASUREData().
This issue also affects measure names, in that context, that start
with I, N, P, V, Y, W or Z.)

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1330