Add .options parser model_binning=false
This is related to issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/144

Which is about making model binning enabled by default.  In nightly
regression it looks like this test, which is faling on every platform,
got broken by this change.

To fix it, for now I'm turning binning off for this test.   That should
get it back to passing.  But, this really *should* pass with it on.  The
main problem seems to be that the L,W parameters are being modified via
global parameters, which means that they can change during the run (such
as with a .STEP command).  But, binning is done in the parser right
now.    So that means that if L,W change, the "bin" that was identified
for them will be potentially wrong.

So this seems to be one use case where the binning code, as it is currently
written, doesn't work.