Remove .options parser model_binning=false
Related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/144

Yesterday, I added the above .options command so that this test would
pass.  This was the one test that failed when I turned on model binning
by default.

I just pushed a code change that fixes this issue so that the above
command is not necessary, so removing it from the netlist.