Test -remeasure with Only .FFT Lines in Netlist
This commit adds a new test for the case of -remeasure on a netlist
which has .FFT lines, but no .MEASURE lines.  Various other .sh
files or Gold Standard files were also updated, to account for
the modified "header text" sent to stdout by the -remeasure
operation.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327