Binning with global params
Related to this issue:

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/144

This test is identical to the comparatorBinning test, except that the
L,W parameters in each BSIM3 MOSFET instance are set with global
parameters.   I recently updated the binning code to allow this.

Note, like the comparatorBinning test, this is a somewhat limited test
in that all the .MODELS are identical.  ie, each "bin" is the same.  So,
this netlist really just tests parsing.  But that is a valid test and
enough for now.