Add Tests for FMIN and FMAX Qualifiers for .FFT Lines
This commit adds subtests for the effects of the FMIN and FMAX
qualifiers, on .FFT lines, on the ENOB, SFDR, SNDR, SNR and THD
metrics.  The comments associated with each subtest line indicate
which metrics and combination of FREQ/FMIN/FMAX settings are "most
tested" by each line.  It also adds various errorexit tests for
"not useful" combinations of FREQ/FMIN/FMAX values that are
considered to be parsing errors in Xyce.

SON Bugs 1327 and 1330
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1330