Create tests for the DELAY device
Issue Xyce#102 and ancient bugzilla issue 165 on joseki (long ago
closed as WONTFIX) called for implementation of a simple delay element
that does nothing more than take its input signal and produce a time-delayed
version of that signal on its output.

The original decision to close joseki 165 as WONTFIX is that it is
fairly simple to use the lossless transmission line as a simple delay
element, and so such a new device was somewhat pointless.

However, recent work exposed issues with the transmission line that
turned out to be fairly deep problems with the transient analysis
package and its handling of dynamic breakpoints.  At first it seemed
that it was worthwhile first to experiment with how the device
performed its interpolation, and ultimately so many changes were
suggested that they seemed inappropriate for the transmission
line.

The delay device was implemented to give us a free hand to try all
those other ideas before it was finally discovered that the problem
really was much deeper than the device itself.  Those have been fixed
in various issues to be enumerated below.

Now we have a usable delay device *and* a customer that is using it,
so it's time to get it into testing.

This commit adds tests of the delay device and all its options.  In
each case, a delay device is fed with a pulsed signal and the netlist
contains an identical pulsed source with a "TD" value --- the output
of the delay device and this second pulsed source should be identical,
and this is tested by outputting the difference between the two
signals and comparing them to a "gold standard" that has all zeros in
that column.

Tested are:
   - the default operation of the device, with quadratic interpolation
     from accepted timesteps in history except immediately after
     detected discontinuities, when linear interpolation is used.  The
     time delay of the device is much longer than the pulse width but
     shorter than the pulse period.

   - The first circuit, but with linear interpolation throughout
     (delay_linear.cir)

   - The first circuit, but using the "current" unconverged solution as
     the third point in the quadratic interpolation scheme (this means
     that interpolation is recomputed for every Newton iteration).
     (delay_noextrap.cir)

   - The first circuit, with breakpoints turned off.  In this case,
     the device still sets maximum timestep equal to the time delay
     when it detects a discontinuity, and relies on the "isConverged"
     technique to prevent acceptance of steps larger than the time
     delay after a discontinuity on the input.
     (delay_nobp.cir)

   - A circuit similar to the first, but with a delay time much
     shorter than the pulse width (in early days, Xyce had a lot of
     trouble with this because it liked to take time steps larger than
     the line delay and resulted in bad interpolation/extrapolation
     artifacts, and exposed the problems with the handling of dynamic
     breakpointing)  (delay_veryshort.cir)

   - A second "veryshort" circuit, but with breakpointing turned
     off (delay_veryshort_nobp.cir)

All tests have gold standards that are generated by perl scripts that
replace the {V(4)-V(2)} output with zero.

Tests Xyce#102  (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/102)
Impacted by Xyce#122 (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/122)
See also Xyce#103 (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/103)
See also Xyce#143 (https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/143)