Add test of external output prepending TIME or FREQ
The external outputter was intended to be designed such that if a user
requests a transient or AC/HB run and specifies an external outputter
that doesn't have TIME or FREQ (respectively) in its output field
list, then we automatically prepend the appropriate independent
variable to the list of requested outputs.  This test checks if that
output appears *anywhere* in the user's requested list, and only
prepends if it is not found at all.

This feature was never tested, and true to form because it wasn't
tested, it didn't work.  There was a missing call to the function that
actually did the work.

Now there is such a call, and this commit adds two tests of it to
existing tests of the external output capability.  This commit also
requires a modification to the testGenCoup test harness in the Xyce repo.
In particular, the "-iotest1" and "-iotest3" options have been
expanded to generate "noindex" output files that do not explicitly
include the independent variable in the output list.

In genext.cir.sh/rlc_series_vccs.cir, we add a noindex .print line in
addition to other existing print lines, and compare the no-index
file to the -iotest1's third output file using file_compare.  They
would not have matched before Xyce#153 (github issue 21) was fixed
because the TIME column wouldn't have been added, and they do now.

In genextAC.cir.sh/rlc_series_vccs_AC.cir we do pretty much the same
thing for an AC circuit, checking that FREQ is appropriately added.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/153
https://github.com/Xyce/Xyce/issues/21