Add bug number for ancient test case
The test case in POLYG dates back to 2005 and claims to test a fix of
a bug in G source POLY statements.  But it predates any attempt by the
Xyce team to make commit messages very meaningful and associate them
carefully with the bug reports that generated the changes.

I hunted down the actual morannon issue associated with this test and
have added it to the test case comments (and to this commit message).

https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=696