Test fixes to error manager static data pollution
The error manager keeps static data and so each instantiation of a
simulator object (either the main API object of class
Xyce::Circuit::Simulator or any of its derived classes) wind up
sharing error information.

This is really bad if there exist more than one object that is active
at a time, and that was not fixed at all.

But it is also really bad if one tries to write a program that
repeatedly allocates simulator objects and run them --- errors in any
run will cause all subsequent attempts to fail because the errors are
not cleaned up in between.  This use case is the subject of Github
issue 19 and gitlab-ex issue Xyce#150.  The open source user was
attempting to write a sort of "Xyce server" that one could submit jobs
to, and was doing it using the general external interface (which he is
using to get access to the general external I/O system so his server
can capture output directly instead of reading output files).

The issue was fixed by making Simulator objects (and by extension any
of objects of their derived classes) call a function to clear the
error counts kept by the error manager.  This test case makes sure it
doesn't get rebugged.

It runs a test harness in the Xyce source tree that does two
allocations of a GenCouplingSimulator object --- one that is
initialized with a known bad netlist that should cause the initialize
step to fail, and a second with a known good netlist (the same as the
bad one, with the error fixed).

The script scans the standard out from the test harness and checks
that the first netlist failed and the second did not, and that the
output contains evidence that the second actually worked as opposed to
just not crashing.

https://github.com/Xyce/Xyce/issues/19
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/150