Add "nompi" tag to a test that cannot work in PbSr
The new test "TestGH19" in the GenCoupAPI directory cannot run when
Xyce libraries have been compiled for parallel.  That is because a
basic operation of the Simulator class (and all its derived classes)
is that the destructor causes MPI_FINALIZE to be called (indirectly,
when the Comm object associated with it is destroyed).

There is basically no quick way to make this test work in parallel
build/serial run mode --- if Xyce is compiled with MPI then the test
simply fails.  On OS X it segfaults, on RHEL7 it crashes with a fatal
error that the second instantiation of the Simulator object has called
MPI functions *after* an MPI_Finalize call, which is illegal.

I'm not even sure if there's a slow way to make that work.  If you
call mpi_finalize it's supposed to be the end of the story, you're not
then supposed to fire up and try to go on.  So this use case may never
work with a parallel build of Xyce, even run serially.  Or at least it
won't work anytime soon, and the test cannot be run this way.

Problem is that Parallel Build/SerialRun testing of Xyce runs all
"serial" tests in the test suite with a binary that was meant for
parallel.  So simply tagging a test as "serial" and not "parallel"
doesn't exclude it from being run this way.

I am therefore adding a tag "nompi" to the TestGH19 test *and* adding
"-nompi" to the tags suggested for PbSr by the suggestXyceTagList.sh
script.  Any test that cannot be run with a parallel build, even
serially, should get the "nompi" tag.

A similar hack has been put into our Jenkins scripts to prevent this
test being run in nightly/weekly/continuous testing.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/150
https://github.com/Xyce/Xyce/issues/19