Replace three raw file gold standards
For Xyce#8 in commit Xyce@2a03db678, I fixed "override" rawfile to
stop indiscriminately replacing the last underscore in an internal
solution variable name with an octothorpe when writing the rawfile
header.  That substitution should never have been happening that way,
it should only have happened for voltage source branch currents ---
that is required for IC-CAP, which expects SPICE3F5 type raw files to
have this variable for every voltage source it creates in a netlist.

This broke three tests, all of which had netlists that had underscores
in solution variable names that were not branch currents, and gold
standards that had the munged name with octothorpes instead.

I have replaced the three problematic gold standards.

Xyce#8
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/8