Test for more permissive .param
Issue #157.

These tests demostrate that a .param can be dependent on TEMP, as long
ias it is indirectly thru a .global_param.  It cannot have shallow
dependence but it is allowed to have a "deep" dependence.

Also added a README, as my intention is for this directory
to cover various aspects of .param refactoring, suggested by
SON bugzilla issue 1190.