Improve Testing of FMAX/MAXFREQ for SNR and SFDR
This commit improves the testing of the FMAX qualifier for .FFT
lines and the MAXFREQ qualifier for .MEASURE FFT lines for the
SNR and SFDR metrics/measures.  It focuses on the "corner case"
where FMAX or MAXFREQ is given but is less than the first harmonic
frequency.

SON Bugs 1280 and 1327
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1280
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1327