Add a directory of Verilog-A plugin tests
For a long time we have not tried to test Verilog-A plugins, because
our regular builds are not shared-library builds with this capability
built in.

But we have been testing shared-library builds in a sort of one-off
Jenkins job for a while now, and I have worked out how to incorporate
tests of the feature into our test framework.

These tests are all tagged with "required:buildplugin" and
"required:installed" because they can only work when Xyce
is *installed* (via make install) and buildxyceplugin exists in the
installation (which only happens if shared libraries are enabled *and*
we're installed).

I have modified "suggestXyceTagList.sh" to try to make the
determination that we're working out of a properly installed build (or
even a kludgy old-school install like in normal testing):

    - it checks whether we're running out of a "bin" directory instead of a
      "src" directory.  If we are, we add "-library" to the tag list.
      - if the directory is "bin" then it also checks if
         buildxyceplugin is present.  If it is, we also add "?buildplugin"
    - it checks whether the lib and include directories exist, and if
      so whether include directory has N_CIR_Xyce.h in it, and if the
      lib directory contains "libxyce.so" and/or libxyce.a.  If it has
      both the header and one of the libraries we addd "?installed" to
      the tag list, because it was installed via "make install".  If
      the shared library is there we also add "?shared" to the tag
      list.

Technically, only the buildplugin really is required here, because
that only gets installed when Xyce is installed properly, but by
adding this extra test we make sure that all the stuff buildxyceplugin
needs to work also got installed.

By changing this way, nightly and CI testing via the normal jobs
should be unchanged, because none of those work via "make install" and
only install via a kludged procedure that dates back to Xyce 1.0.
Furthermore, they don't (yet) use "suggestXyceTagList.sh" and so won't
ever try to run these new tests with their required tags.

The one-off Jenkins jobs that test shared builds, however, should
automagically start running these tests as soon as they pull the test
repo.